start of work to pair flavor with a model
chantilly has a global metrics and stats, and then when a new flavor is added the old ones
are dropped. Instead, this method will require a model flavor when adding a model,
and that model name will be stored with the index for stats and metrics (and flavor)
so we can have more than one model going at the same time. We also never automatically drop
the database.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>