adding label endpoint and multiclass example
since there is functionality shared between learn and label, I have refactored
the client functions to share that. Also, the check_model function is not useful
because of the design of the river classes they always have the function attributes.
I have left this for now. Instead of hard coding one predict function (e.g.,
predict_proba_one vs predict_one) given that some models can be used in
different contexts, we instead have a list of methods to try.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>